pdf-chunk-func/requirements.txt
azure-functions
azure-search-documents
PyMuPDF
pillow
pytesseract
tiktoken
azure-core
python-dotenv
openai